fix(reservation): null-safe exception handler, safe cast, race fix, tests
- Guard GlobalExceptionHandler Map.of against null getMessage() via Objects.toString
- Replace unsafe (int) cast with Math.toIntExact for guest_count overflow detection
- Fix docker-compose race: spanner-init-reservation depends_on spanner-init
  with condition: service_completed_successfully
- Add idempotency test: verify second createWithSlotValidation throws ALREADY_EXISTS
  and no duplicate row is created
- Add controller test: verify idempotent replay returns 200 OK (not 201)

+++ services/reservation-service/src/main/java/com/asoviewclone/reservation/repository/ReservationRepository.java
@@ -595,7 +595,7 @@ public class ReservationRepository {
         rs.getString("idempotency_key"),
         rs.getString("guest_name"),
         rs.getString("guest_email"),
-        (int) rs.getLong("guest_count"),
+        Math.toIntExact(rs.getLong("guest_count")),
         rs.isNull("reject_reason") ? null : rs.getString("reject_reason"),
         rs.isNull("cancel_reason") ? null : rs.getString("cancel_reason"),
         rs.getTimestamp("created_at").toSqlTimestamp().toInstant(),
